requirements.txt
numpy>=1.22.0
pandas>=1.4.0
ccxt>=2.0.0
python-telegram-bot>=13.0
matplotlib>=3.5.0
ta>=0.10.0
schedule>=1.1.0
python-dotenv>=0.19.0
requests>=2.26.0
